requirements.txt
streamlit>=1.30.0
llama-cpp-python>=0.2.30
gdown>=5.1.0
ninja
